Adding unit tests and e2e tests
