+++ server/events/project_command_builder_internal_test.go
@@ -640,6 +640,203 @@ projects:
 	}
 }
 
+func TestBuildProjectCmdCtx_WithPolicCheckEnabled(t *testing.T) {
+	emptyPolicySets := models.PolicySets{
+		Version:    nil,
+		PolicySets: []models.PolicySet{},
+	}
+	baseRepo := models.Repo{
+		FullName: "owner/repo",
+		VCSHost: models.VCSHost{
+			Hostname: "github.com",
+		},
+	}
+	pull := models.PullRequest{
+		BaseRepo: baseRepo,
+	}
+	cases := map[string]struct {
+		globalCfg           string
+		repoCfg             string
+		expErr              string
+		expCtx              models.ProjectCommandContext
+		expPolicyCheckSteps []string
+	}{
+		// Test that if we've set global defaults and no project config
+		// that the global defaults are used.
+		"global defaults": {
+			globalCfg: `
+repos:
+- id: /.*/
+`,
+			repoCfg: "",
+			expCtx: models.ProjectCommandContext{
+				ApplyCmd:           "atlantis apply -d project1 -w myworkspace",
+				BaseRepo:           baseRepo,
+				EscapedCommentArgs: []string{`\f\l\a\g`},
+				AutomergeEnabled:   false,
+				AutoplanEnabled:    true,
+				HeadRepo:           models.Repo{},
+				Log:                nil,
+				PullMergeable:      true,
+				Pull:               pull,
+				ProjectName:        "",
+				ApplyRequirements:  []string{},
+				RePlanCmd:          "atlantis plan -d project1 -w myworkspace -- flag",
+				RepoRelDir:         "project1",
+				User:               models.User{},
+				Verbose:            true,
+				Workspace:          "myworkspace",
+				PolicySets:         emptyPolicySets,
+			},
+			expPolicyCheckSteps: []string{"show", "policy_check"},
+		},
+
+		// If the repos are allowed to set everything then their config should
+		// come through.
+		"full repo permissions": {
+			globalCfg: `
+repos:
+- id: /.*/
+  workflow: default
+  apply_requirements: [approved]
+  allowed_overrides: [apply_requirements, workflow]
+  allow_custom_workflows: true
+workflows:
+  default:
+    policy_check:
+      steps: []
+`,
+			repoCfg: `
+version: 3
+automerge: true
+projects:
+- dir: project1
+  workspace: myworkspace
+  autoplan:
+    enabled: true
+    when_modified: [../modules/**/*.tf]
+  terraform_version: v10.0
+  apply_requirements: []
+  workflow: custom
+workflows:
+  custom:
+    policy_check:
+      steps:
+      - policy_check
+`,
+			expCtx: models.ProjectCommandContext{
+				ApplyCmd:           "atlantis apply -d project1 -w myworkspace",
+				BaseRepo:           baseRepo,
+				EscapedCommentArgs: []string{`\f\l\a\g`},
+				AutomergeEnabled:   true,
+				AutoplanEnabled:    true,
+				HeadRepo:           models.Repo{},
+				Log:                nil,
+				PullMergeable:      true,
+				Pull:               pull,
+				ProjectName:        "",
+				ApplyRequirements:  []string{},
+				RepoConfigVersion:  3,
+				RePlanCmd:          "atlantis plan -d project1 -w myworkspace -- flag",
+				RepoRelDir:         "project1",
+				TerraformVersion:   mustVersion("10.0"),
+				User:               models.User{},
+				Verbose:            true,
+				Workspace:          "myworkspace",
+				PolicySets:         emptyPolicySets,
+			},
+			expPolicyCheckSteps: []string{"policy_check"},
+		},
+	}
+
+	for name, c := range cases {
+		t.Run(name, func(t *testing.T) {
+			tmp, cleanup := DirStructure(t, map[string]interface{}{
+				"project1": map[string]interface{}{
+					"main.tf": nil,
+				},
+				"modules": map[string]interface{}{
+					"module": map[string]interface{}{
+						"main.tf": nil,
+					},
+				},
+			})
+			defer cleanup()
+
+			workingDir := NewMockWorkingDir()
+			When(workingDir.Clone(matchers.AnyPtrToLoggingSimpleLogger(), matchers.AnyModelsRepo(), matchers.AnyModelsPullRequest(), AnyString())).ThenReturn(tmp, false, nil)
+			vcsClient := vcsmocks.NewMockClient()
+			When(vcsClient.GetModifiedFiles(matchers.AnyModelsRepo(), matchers.AnyModelsPullRequest())).ThenReturn([]string{"modules/module/main.tf"}, nil)
+
+			// Write and parse the global config file.
+			globalCfgPath := filepath.Join(tmp, "global.yaml")
+			Ok(t, ioutil.WriteFile(globalCfgPath, []byte(c.globalCfg), 0600))
+			parser := &yaml.ParserValidator{}
+			globalCfg, err := parser.ParseGlobalCfg(globalCfgPath, valid.NewGlobalCfg(false, false, false))
+			Ok(t, err)
+
+			if c.repoCfg != "" {
+				Ok(t, ioutil.WriteFile(filepath.Join(tmp, "atlantis.yaml"), []byte(c.repoCfg), 0600))
+			}
+
+			builder := NewProjectCommandBuilder(
+				true,
+				parser,
+				&DefaultProjectFinder{},
+				vcsClient,
+				workingDir,
+				NewDefaultWorkingDirLocker(),
+				globalCfg,
+				&DefaultPendingPlanFinder{},
+				&CommentParser{},
+				false,
+			)
+
+			cmd := models.PolicyCheckCommand
+			t.Run(cmd.String(), func(t *testing.T) {
+				ctxs, err := builder.buildProjectCommandCtx(&CommandContext{
+					Pull: models.PullRequest{
+						BaseRepo: baseRepo,
+					},
+					PullMergeable: true,
+				}, models.PlanCommand, "", []string{"flag"}, tmp, "project1", "myworkspace", true)
+
+				if c.expErr != "" {
+					ErrEquals(t, c.expErr, err)
+					return
+				}
+
+				ctx := ctxs[1]
+
+				Ok(t, err)
+
+				// Construct expected steps.
+				var stepNames []string
+				var expSteps []valid.Step
+
+				stepNames = c.expPolicyCheckSteps
+				for _, stepName := range stepNames {
+					expSteps = append(expSteps, valid.Step{
+						StepName: stepName,
+					})
+				}
+
+				c.expCtx.CommandName = cmd
+				// Init fields we couldn't in our cases map.
+				c.expCtx.Steps = expSteps
+				ctx.PolicySets = emptyPolicySets
+
+				Equals(t, c.expCtx, ctx)
+				// Equals() doesn't compare TF version properly so have to
+				// use .String().
+				if c.expCtx.TerraformVersion != nil {
+					Equals(t, c.expCtx.TerraformVersion.String(), ctx.TerraformVersion.String())
+				}
+			})
+		})
+	}
+}
+
 func mustVersion(v string) *version.Version {
 	vers, err := version.NewVersion(v)
 	if err != nil {
