+++ server/testfixtures/test-repos/simple/exp-output-atlantis-policy-check.txt
@@ -1,0 +1,1 @@
+no template matched–this is a bug
